refactor: Loadingコンポーネントをcomponentsディレクトリに移動
